Refactored frontend code
