+++ ui/src/app.tsx
@@ -1,9 +1,11 @@
-import { match } from 'assert';
 import React, { useEffect, useState } from 'react';
-import { Router, Routes, Route, BrowserRouter, Navigate } from 'react-router-dom';
-import api from './api';
+import { Routes, Route, BrowserRouter, Navigate } from 'react-router-dom';
 import { Chats } from './components/Chats/Chats';
+import { ChatsConversation } from './components/Chats/ChatsConversation/ChatsConversation';
+import { DeleteModal } from './components/Details/DeleteModal/DeleteModal';
 import { Details } from './components/Details/Details';
+import { DetailsConversation } from './components/Details/DetailsConversation/DetailsConversation';
+import { EditModal } from './components/Details/EditModal/EditModal';
 import { List } from './components/List/List';
 import { Menubar } from './components/Menubar/Menubar';
 import { PublishAd } from './components/PublishAd/PublishAd';
@@ -12,13 +14,12 @@ import { useAppDispatch, useAppSelector } from './redux/hooks/hooks';
 import { loadState } from './redux/slices/classifiedsSlice';
 
 export function App() {
-  const dispatch = useAppDispatch();
   const { ads, myads } = useAppSelector((state) => state.classifieds.data);
+  const dispatch = useAppDispatch();
 
+  // TODO: unsubscribe on unmount
   useEffect(() => {
     dispatch(loadState());
-    document.body.style.overflowX = "hidden";
-    document.body.style.overflowY = "auto";
   }, []);
 
   return (
@@ -26,23 +27,26 @@ export function App() {
       <Menubar />
       <UserInfo />
       <Routes >
-        <Route path="/myads/:id" element={<Details />} />
-        <Route path="/ads/:id" element={<Details />} />
-        <Route path="*" element={<></>} />
-      </Routes>
-      <Routes >
-        <Route path="/ads" >
-          <Route path="*" index element={<List listAds={ads} />} />
-          {/* <Route path="page/:page" element={<List listAds={ads} />} /> */}
+        <Route path="/ads/*" element={<List listAds={ads} />} />
+        <Route path="/myads/*" element={<List listAds={myads} />} />
+        <Route path="/chat" element={<Chats />}>
+          <Route path=":id" element={<ChatsConversation />} />
         </Route>
-        <Route path="/myads" >
-          <Route path="*" index element={<List listAds={myads} />} />
-          {/* <Route path="page/:page" element={<List listAds={myads} />} /> */}
-        </Route>
-        <Route path="/chat" element={<Chats />} />
         <Route path="/newad" element={<PublishAd />} />
         <Route path="*" element={<Navigate to="/ads" replace />} />
       </Routes>
+      <Routes >
+        <Route path="ads/:id" element={<Details />}>
+          <Route path="chat" element={<DetailsConversation />} />
+        </Route>
+        <Route path="myads/:id" element={<Details />}>
+          <Route path="delete" element={<DeleteModal />} />
+          <Route path="edit" element={<EditModal />} />
+        </Route>
+        <Route path="*" element={<></>} />
+      </Routes>
+
+
     </BrowserRouter>
   );
 }
