+++ ui/src/components/Details/DeleteModal/style.ts
