+++ ui/src/components/Details/EditModal/style.ts
