+++ ui/src/components/Details/ForwardModal/style.ts
