+++ ui/src/components/Details/style.ts
@@ -1,7 +1,7 @@
-import { RiDeleteBinLine, RiShareForwardLine, RiStarLine, RiCloseFill, RiSendPlane2Fill, RiSendPlane2Line, RiStarFill, RiEditLine, RiPriceTag3Line } from 'react-icons/ri';
+import { RiDeleteBinLine, RiShareForwardLine, RiCloseFill, RiSendPlane2Line, RiEditLine, RiPriceTag3Line } from 'react-icons/ri';
 import { BiChat } from 'react-icons/bi';
-import styled, { keyframes } from 'styled-components';
-import { MdNavigateNext, MdOutlineArrowBackIosNew, MdOutlineArrowLeft, MdOutlineArrowRight } from 'react-icons/md';
+import styled from 'styled-components';
+import { MdNavigateNext, MdOutlineArrowBackIosNew } from 'react-icons/md';
 
 export const DetailsBackground = styled.div`
     width: 100%;
@@ -85,8 +85,6 @@ export const SmallScreenTitle = styled.div`
         display: none;
     }
 `;
-
-
 
 export const Tags = styled.div`
     display: flex;
@@ -237,153 +235,6 @@ export const PriceIcon = styled(RiPriceTag3Line)`
    margin-top: -.12em;
 `;
 
-export const Images = styled.div`
-    display: flex;
-    flex-direction: column;
-    width: 50%;
-    height: 30em; 
-    @media (max-width: 600px) {
-        width: 100%;
-        height: 15em;
-    }
-`;
-
-export const ImageWrapper = styled.div`
-    min-width: 100%;
-    max-width: 100%;
-    position: relative;
-    -webkit-user-select: none;
-    -ms-user-select: none; 
-    user-select: none; 
-`;
-
-interface ArrowProps {
-    $toShow: boolean,
-    $disabled: boolean,
-}
-
-export const RightArrow = styled(MdOutlineArrowRight) <ArrowProps>`
-    position: absolute;
-    right: 0;
-    top: 50%;
-    transform: translate(0, -50%); 
-    font-size: 2.5rem;
-    cursor: ${p => p.$disabled ? 'default' : 'pointer'};
-    color: ${p => p.$disabled ? 'rgba(0,0,0,.1)' : 'rgba(0,0,0,.8)'};
-    background:  rgba(236,243,243,.5);
-    border-radius: .25rem;
-    visibility: ${p => p.$toShow ? 'visible' : 'hidden'};
-
-    @media (max-width: 600px) {
-        visibility: visible;
-        font-size: 2rem;
-    }
-`;
-
-export const LeftArrow = styled(MdOutlineArrowLeft) <ArrowProps>`
-    position: absolute;
-    left: 0;
-    top: 50%;
-    font-size: 2.5rem;
-    transform: translateY(-50%); 
-    cursor: ${p => p.$disabled ? 'default' : 'pointer'};
-    color: ${p => p.$disabled ? 'rgba(0,0,0,.1)' : 'rgba(0,0,0,.8)'};
-    background:  rgba(236,243,243,.5);
-    visibility: ${p => p.$toShow ? 'visible' : 'hidden'};
-    border-radius: .25rem;
-
-    @media (max-width: 600px) {
-        visibility: visible;
-        font-size: 2rem;
-    }
-
-`;
-
-export const Image = styled.img`
-    border-radius: .25rem;
-    object-fit: contain;
-    box-shadow: rgba(0, 0, 0, 0.15) 0px 5px 15px 0px;
-    border: 1px solid rgba(0,0,0,.2);
-    min-height: 20em;
-    max-height: 20em;
-    width: 100%;
-    -webkit-user-select: none;
-    -ms-user-select: none; 
-    user-select: none; 
-    @media (max-width: 600px) {
-        min-height: 15em;
-        max-height: 15em;
-    }
-
-`;
-
-export const SecondaryImages = styled.div`
-    display: flex;
-    z-index: 110;
-    width: 100%;
-    height: 15em;
-    align-items: center;
-    overflow-y: hidden;
-    overflow-x: auto;
-    scroll-behavior: smooth;    
-    border-radius: .25rem;
-    padding: 0 .1em;
-
-    @media (max-width: 600px) {
-        display: none;
-    };
-
-    ::-webkit-scrollbar {
-      width: 10px;
-      height: 10px;
-    }
-    ::-webkit-scrollbar-track {
-      border: 1px solid rgba(0,0,0,.2);
-      background: rgba(253,253,253,1);
-      border-radius: .25rem; 
-    }
-    ::-webkit-scrollbar-thumb {
-      border-radius: .25rem;
-      background: white;
-      box-shadow: inset 0 0 1px 1px rgba(0,0,0,.6);
-    }
-`;
-
-interface SecondaryImagesProps {
-    isSelected: boolean,
-
-}
-export const SecondaryImage = styled.img<SecondaryImagesProps>`
-    border-radius: .25rem;
-    object-fit: contain;
-    cursor: pointer;
-    &:hover {
-        border-style: double;
-        border-color: black;
-        border-width: medium;
-    }
-    border-style: ${p => p.isSelected ? 'double' : 'none;'};
-    border-color: ${p => p.isSelected ? 'black' : 'none;'};
-    border-width: ${p => p.isSelected ? 'thick' : 'none;'};
-
-    background: white;
-    box-shadow: rgba(50, 50, 93, 0.25) 0px 2px 5px -1px, rgba(0, 0, 0, 0.3) 0px 1px 3px -1px;
-
-    margin-right: 1em;
-    &:last-child{
-        margin-right: 0;
-    }
-    max-width: 7em;
-    min-width: 7em;
-    max-height: 7em;
-    min-height: 7em;
-
-    @media (max-width: 600px) {
-        margin-right: .8em;
-        margin-bottom: 0;
-    }
-`;
-
 export const InfoColumn = styled.div`
     display: flex;  
     flex-direction: column;
@@ -451,27 +302,6 @@ export const ForwardButton = styled.button`
     z-index: 110;
 `;
 
-interface FavProps {
-    isFavorited: boolean,
-}
-
-export const FavButton = styled.button<FavProps>`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 8em;
-    padding: .5em .5em;
-    border-radius: .25rem;
-    font-size: .64rem;
-    font-weight: 700;
-    background: white;
-    box-shadow: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px, rgba(60, 64, 67, 0.15) 0px 1px 3px 1px; 
-    margin-right: 1em;
-    z-index: 110;
-    color: ${p => p.isFavorited ? 'orange' : 'black'};
-
-`;
-
 export const DeleteButton = styled.button`
     display: flex;
     justify-content: center;
@@ -536,18 +366,6 @@ export const EditIcon = styled(RiEditLine)`
     font-size: .9rem;
 `;
 
-export const FavIcon = styled(RiStarLine)`
-     font-size: .9rem;
-    margin-right: .3em;
-`;
-
-
-export const FavIconClicked = styled(RiStarFill)`
-    font-size: .9rem;
-    margin-right: .3em;
-    color: orange;
-`;
-
 export const ChatIcon = styled(BiChat)`
     font-size: .9rem;
     margin-right: .3em;
@@ -564,193 +382,7 @@ export const GoBackIcon = styled(MdOutlineArrowBackIosNew)`
     transform: translateY(-50%);
     position: absolute;
 `;
-export const ConversationWrapper = styled.div`
-    height: 20rem;
-    width: 24em; 
-    position: absolute;
-    right: 0;
-    top: 27em;
-    z-index: 200;
 
-    @media (max-width: 600px) {
-        top: 0;
-        position: fixed;
-    }
-`;
-
-export const Conversation = styled.div`
-    border-radius: .25rem;
-    width: 24em; 
-    box-shadow: rgba(0, 0, 0, 0.15) 0px 5px 15px 0px;
-    height: 18rem;
-    background: white;  
-    @media (max-width: 600px) {
-        z-index: 200;
-        width: 100%; 
-        height: 100%;
-        right: 0;
-        border-radius: 0;
-        box-shadow: none;
-        position: fixed;
-    }
-`;
-
-export const ConversationUpperRow = styled.div`
-    display: flex;
-    flex-direction: column; 
-    align-items: center;
-    justify-content: center;
-    height: 15%;
-    width: 100%;
-`;
-
-export const ConversationBottomRow = styled.div`
-    height: 15%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-`;
-
-export const ConversationReceiver = styled.div`
-    display: flex;
-    align-items: center;
-    height: 1rem;
-    font-size: .7rem;
-`;
-
-export const ConversationReceiverShip = styled.div`
-    margin-left: .5em;
-`;
-
-export const MessageList = styled.div`
-    height: 15rem;
-    display: flex;
-    flex-direction: column;
-    border-bottom: 1px solid rgba(0, 0, 0, .2);
-    border-top: 1px solid rgba(0, 0, 0, .2);
-    padding: 1em .75em;
-    height: 70%;
-    overflow-y: auto;
-    overflow-x: hidden;
-
-    ::-webkit-scrollbar {
-        width: 10px;
-        height: 10px;
-    }
-    ::-webkit-scrollbar-track {
-        border: 1px solid rgba(0,0,0,.2);
-        background: rgba(253,253,253,1);
-        border-top-right-radius: .25rem; 
-        border-bottom-right-radius: .25rem; 
-    }
-    ::-webkit-scrollbar-thumb {
-        border-radius: .25rem;
-        background: white;
-        box-shadow: inset 0 0 1px 1px rgba(0,0,0,.6);
-    }
-`;
-
-export const SentMessage = styled.div`
-    border: 1px solid rgba(0, 0, 0, .2);
-    border-radius: .25rem;
-    max-width: 60%;
-    align-self: end;
-    padding: .5em 1em;
-    margin-bottom: .5rem;
-    background: rgba(0, 0, 0, .015);
-    line-height: 1;
-
-    &:last-child{
-        margin-bottom: 0em;
-    }
-`;
-
-export const SigilContainer = styled.div`
-`;
-
-
-export const ReceivedMessage = styled.div`
-    align-self: start;
-    margin-bottom: .5rem;
-    display: flex;
-    max-width: 60%;
-    line-height: 1;
-
-    &:last-child{
-        margin-bottom: 0;
-    }
-`;
-
-export const ReceivedMessageBox = styled.div`
-    display: flex;
-    flex-direction: column;
-    font-size: .7rem;
-    margin-bottom: .3em;
-    border: 1px solid rgba(0, 0, 0, .2);
-    border-radius: .25rem;
-    background: rgba(0, 0, 0, .015);
-    padding: .5em 1em;
-`;
-
-export const MessageShip = styled.div`  
-    margin-left: .3rem;
-`;
-
-export const MessageText = styled.div`
-    display: flex;
-    font-size: .65rem;
-    line-height: 1;
-    margin-bottom: .3em;
-    word-break: break-all;
-`;
-
-export const MessageDate = styled.div`
-    font-size: .6rem;
-    display: flex;
-    justify-content: end;
-    color: rgba(0, 0, 0, .5);
-    white-space: nowrap;
-`;
-
-export const InputRow = styled.div`
-    height: 3rem;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    position: relative;
-    padding: 0 1em; 
-`;
-
-export const Input = styled.input`
-    width: 100%;
-    border: 1px solid rgba(0, 0, 0, .2);
-    box-shadow: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px, rgba(60, 64, 67, 0.15) 0px 1px 3px 1px;
-    border-radius: .25rem;
-    padding: .3em 2em .3em 1em;
-    font-size: .65rem;
-`;
-
-export const SendIcon = styled(RiSendPlane2Line)`
-    position: absolute;
-    right: 1.5em;
-    cursor: pointer;
-`;
-
-export const ConversationAdTitle = styled.div`
-    font-size: .7rem;
-    padding: 0 1em;
-    font-weight: 900;
-    height: 1rem;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-    width: 100%;
-    text-align: center;
-
-    @media (max-width: 600px) {
-        padding: 0 3.5em;
-    }
-`;
 
 export const NavigatedIcon = styled(MdNavigateNext)`
