+++ ui/src/components/List/Card/Card.tsx
@@ -1,11 +1,10 @@
 import { reactRenderer, sigil } from "@tlon/sigil-js";
-import moment from "moment";
 import React from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import api from "../../api";
-import { useAppSelector } from "../../redux/hooks/hooks";
-import { Advertisement } from "../../types";
-import { daToDate, locationHasPath } from "../../util";
+import api from "../../../api";
+import { useAppSelector } from "../../../redux/hooks/hooks";
+import { Advertisement } from "../../../types";
+import { daToDate } from "../../../util";
 import { CardContainer, Image, Publisher, Title, Date, ForwardIcon, BottomRow, DeleteIcon, FavIcon, Price, ContentContainer, MiddleRow, Icons, PublisherInfo, PriceLabel, PriceContainer, SourceContainer, PriceIcon } from './style';
 
 interface CardProps {
@@ -13,16 +12,14 @@ interface CardProps {
 };
 
 export const Card = ({ advertisement }: CardProps) => {
+
     const formatedDate = daToDate(advertisement!.date!).fromNow();
     const favorites = useAppSelector((state) => state.classifieds.data.favorites);
     const navigate = useNavigate();
-    let location = useLocation();
 
     const handleCardClick = () => {
-        location.pathname.indexOf('/ads') > -1 ?
-            navigate(`/ads/${advertisement.id}`) :
-            navigate(`/myads/${advertisement.id}`)
-    }
+        navigate(`${advertisement.id}`);
+    };
 
     return (
         <CardContainer onClick={handleCardClick}>
@@ -32,7 +29,8 @@ export const Card = ({ advertisement }: CardProps) => {
                 <Image src='/apps/classifieds/assets/placeholder.png' />
             }
             <ContentContainer>
-                <Title>{advertisement?.title}
+                <Title>
+                    {advertisement?.title}
                     {favorites?.includes(advertisement.id) && < FavIcon />}
                     {/* <ForwardIcon /> */}
                 </Title>
