+++ ui/src/components/Menubar/Menubar.tsx
@@ -8,10 +8,10 @@ export const Menubar = () => {
         <MenubarContainer>
             <Title>Classifieds</Title>
             <Tabs>
-                <Tab to="ads" ><TheiAdsIcon /><TabText>Ads</TabText></Tab>
-                <Tab to="myads" ><MyAdsIcon /><TabText>My Ads</TabText></Tab>
-                <Tab to="chat" ><ChatIcon /><TabText>Chat</TabText></Tab>
-                <Tab to="newad" ><PublishAdIcon /><TabText>New Ad</TabText></Tab>
+                <Tab to="/ads" ><TheiAdsIcon /><TabText>Ads</TabText></Tab>
+                <Tab to="/myads" ><MyAdsIcon /><TabText>My Ads</TabText></Tab>
+                <Tab to="/chat" ><ChatIcon /><TabText>Chat</TabText></Tab>
+                <Tab to="/newad" ><PublishAdIcon /><TabText>New Ad</TabText></Tab>
             </Tabs>
         </MenubarContainer>
     )
