+++ ui/vite.config.ts
@@ -5,7 +5,6 @@ import { urbitPlugin } from '@urbit/vite-plugin-urbit';
 export default ({ mode }) => {
   Object.assign(process.env, loadEnv(mode, process.cwd()));
   const SHIP_URL = process.env.SHIP_URL || process.env.VITE_SHIP_URL || 'http://localhost:8080';
-  console.log(SHIP_URL);
 
   return defineConfig({
     plugins: [urbitPlugin({ base: 'classifieds', target: SHIP_URL, secure: false }), reactRefresh()],
